requirements.txt
# Core dependencies
pandas==2.2.3
openpyxl==3.1.5
numpy==2.2.4
pillow==11.1.0
psutil==7.0.0

# OCR and PDF processing
easyocr==1.7.2
PyMuPDF==1.25.5
pdf2image==1.17.0
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0

# Google Cloud dependencies
google-cloud-documentai==3.3.0
google-genai==1.9.0
google-oauth2-tool==0.0.3

# Time zone support
tzdata==2025.1

# Logging and utilities
python-json-logger==3.3.0
rich==14.0.0
alive-progress==3.2.0
python-dotenv==1.1.0

# another dependencies
about-time==4.2.1
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
filelock==3.18.0
fsspec==2025.3.0
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
grapheme==0.6.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.3
oauth2client==4.1.3
opencv-python-headless==4.11.0.86
packaging==24.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
PyPDF2==3.0.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
scikit-image==0.25.2
scipy==1.15.2
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
tifffile==2025.3.13
ttkthemes==3.2.2
typing-inspection==0.4.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uuid==1.30
websockets==15.0.1
